pyproject.toml
[tool.poetry]
name = "denoise"
version = "0.1.0"
description = ""
authors = ["Daniel Nobbe <daniellnobbe@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.25.2"
lightning = "2.1.0"
torchvision = "^0.15.2"
tensorboardx = "^2.6.2.2"
tensorboard = "^2.14.1"
kornia = {path = "../kornia"}
streamlit = "^1.28.1"
imageio = "^2.32.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
